requirements.txt
requests
python-dotenv
pandas
openpyxl
msal
fastapi
uvicorn
